+++ plugin.c
@@ -44,6 +44,9 @@ EXTERN_MODULE(label);
 EXTERN_MODULE(mpd);
 #endif
 EXTERN_MODULE(network);
+#if defined(PLUGIN_ENABLED_PULSE)
+EXTERN_MODULE(pulse);
+#endif
 EXTERN_MODULE(removables);
 EXTERN_MODULE(river);
 EXTERN_MODULE(sway_xkb);
@@ -129,6 +132,9 @@ init(void)
     REGISTER_CORE_MODULE(mpd, mpd);
 #endif
     REGISTER_CORE_MODULE(network, network);
+#if defined(PLUGIN_ENABLED_PULSE)
+    REGISTER_CORE_MODULE(pulse, pulse);
+#endif
     REGISTER_CORE_MODULE(removables, removables);
 #if defined(HAVE_PLUGIN_river)
     REGISTER_CORE_MODULE(river, river);